Added functions to compute point-to-point variance, doubling time, e-folding time. Added wrapper functions for fractional variance, point-to-point variance, doubling time, e-folding time to be used with lightcurve Flux Points objects.
