+++ gammapy/estimators/utils.py
@@ -7,6 +7,12 @@ from astropy.table import Table
 from gammapy.datasets import SpectrumDataset, SpectrumDatasetOnOff
 from gammapy.datasets.map import MapEvaluator
 from gammapy.maps import WcsNDMap
+from gammapy.stats import (
+    compute_fvar,
+    compute_fpp,
+    compute_etime,
+    compute_2time,
+)
 from gammapy.modeling.models import (
     ConstantFluxSpatialModel,
     PowerLawSpectralModel,
@@ -17,6 +23,9 @@ __all__ = [
     "estimate_exposure_reco_energy",
     "find_peaks",
     "resample_energy_edges",
+    "lc_fvar",
+    "lc_fpp",
+    "eval_lc_timing",
 ]
 
 
@@ -226,3 +235,73 @@ def resample_energy_edges(dataset, conditions={}):
                 energy_edges.append(energy_min)
                 break
     return u.Quantity(energy_edges[::-1])
+
+
+def lc_fvar(lightcurve):
+    r"""Wrapper to utilize the '~gammapy.stats.compute_fvar' function
+    directly from the lightcurve FluxPoints object
+
+    Parameters
+    ----------
+    lightcurve : '~gammapy.estimators.FluxPoints'
+        the lightcurve object
+
+
+    Returns
+    -------
+    fvar, fvar_err : `~numpy.ndarray`
+        Fractional excess variance.
+    """
+
+    flux = lightcurve.flux.data.flatten()
+    flux_err = lightcurve.flux_err.data.flatten()
+
+    return compute_fvar(flux, flux_err)
+
+
+def lc_fpp(lightcurve):
+    r"""Wrapper to utilize the '~gammapy.stats.compute_fpp' function
+    directly from the lightcurve FluxPoints object
+
+    Parameters
+    ----------
+    lightcurve : '~gammapy.estimators.FluxPoints'
+        the lightcurve object
+
+    Returns
+    -------
+    fpp: `~numpy.ndarray`
+        Point-to-point fractional variation
+
+    """
+
+    flux = lightcurve.flux.data.flatten()
+    flux_err = lightcurve.flux_err.data.flatten()
+
+    return compute_fpp(flux, flux_err)
+
+
+def eval_lc_timing(lightcurve, efolding=True):
+    r"""Wrapper to utilize the '~gammapy.stats.compute_etime'
+    and '~gammapy.stats.compute_2time' functions
+    directly from the lightcurve FluxPoints object
+
+    Parameters
+    ----------
+    lightcurve : '~gammapy.estimators.FluxPoints'
+        the lightcurve object
+    efolding : 'bool'
+        flag to compute e-folding time or doubling time
+
+    Returns
+    -------
+    t:  float
+        Doubling time or e-folding time
+    """
+    flux = lightcurve.flux.data.flatten()
+    time = lightcurve.geom.axes["time"].time_mid
+
+    if efolding:
+        return compute_etime(flux, time)
+    else:
+        return compute_2time(flux, time)
