+++ gammapy/stats/variability.py
@@ -67,64 +67,6 @@ def compute_fvar(flux, flux_err):
     return fvar, fvar_err
 
 
-def weighted_fvar(flux, flux_err, time):
-    r"""Calculate the time_weighted fractional excess variance.
-
-    This method accesses the "FLUX", "FLUX_ERR" columns
-    from the lightcurve data, and the "TIME" axis.
-
-    A weighted fractional excess variance implementation
-     useful in the case of non-uniform binning in time
-
-    Obtained from the unweighted computation by substituting weighted averages
-    to the mean flux and mean error on flux :math: "\bar{\sigma^2}" and
-    :math: "\frac{T}{T^2 - \sum{t_i^2}} to the usual Bessel correction :math: "\frac{1}{N - 1}"
-    where T is the sum of the time intervals.
-
-    In the case of uniform time bins this reduces back to the unweighted F_var
-
-    Parameters
-    ----------
-    flux : `~astropy.units.Quantity`
-        the measured fluxes
-    flux_err : `~astropy.units.Quantity`
-        the error on measured fluxes
-    time : `~astropy.units.Quantity`
-        the time bin lengths
-
-    Returns
-    -------
-    fvar, fvar_err : `~numpy.ndarray`
-        Fractional excess variance.
-
-    References
-    ----------
-    .. [Vaughan2003] "On characterizing the variability properties of X-ray light
-       curves from active galaxies", Vaughan et al. (2003)
-       https://ui.adsabs.harvard.edu/abs/2003MNRAS.345.1271V
-    .. [Emery2020] "Study of the variability of active galactic nuclei at very
-       high energy with H.E.S.S.", Emery Gabriel, PhD thesis (2020)
-    .. GNU Scientific Library
-       https://www.gnu.org/software/gsl/doc/html/statistics.html#weighted-samples
-    """
-
-    flux_mean = np.average(flux, weights=time.jd)
-    interval = np.sum(time.jd)
-
-    s_square = (interval / (interval**2 - np.sum(time.jd**2))) * np.sum(
-        time.jd * (flux - flux_mean) ** 2
-    )
-    sig_square = np.nansum(time.jd * flux_err**2) / interval
-    fvar = np.sqrt(np.abs(s_square - sig_square)) / flux_mean
-
-    sigxserr_a = np.sqrt(2 / interval) * sig_square / flux_mean**2
-    sigxserr_b = np.sqrt(sig_square / interval) * (2 * fvar / flux_mean)
-    sigxserr = np.sqrt(sigxserr_a**2 + sigxserr_b**2)
-    fvar_err = sigxserr / (2 * fvar)
-
-    return fvar, fvar_err
-
-
 def compute_fpp(flux, flux_err):
     r"""Calculate the point-to-point fractional variation.
 
@@ -271,82 +213,3 @@ def compute_etime(flux, time):
     te = np.amin(tij)
 
     return te
-
-
-def lc_fvar(lightcurve, weighted=False):
-    r"""Wrapper to utilize the '~gammapy.stats.compute_fvar' function
-    directly from the lightcurve FluxPoints object
-
-    This method uses a (optional) boolean flag to add weights based on bin width
-    useful in cases of non-uniform time binning.
-
-    Parameters
-    ----------
-    lightcurve : '~gammapy.estimators.FluxPoints'
-        the lightcurve object
-    weighted : bool
-        flag to indicate if the normal or weighted F_var needs to be computed
-
-    Returns
-    -------
-    fvar, fvar_err : `~numpy.ndarray`
-        Fractional excess variance.
-    """
-
-    flux = lightcurve.flux.data.flatten()
-    flux_err = lightcurve.flux_err.data.flatten()
-
-    if weighted:
-        time = lightcurve.geom.axes["time"].time_delta
-        return weighted_fvar(flux, flux_err, time)
-
-    else:
-        return compute_fvar(flux, flux_err)
-
-
-def lc_fpp(lightcurve):
-    r"""Wrapper to utilize the '~gammapy.stats.compute_fpp' function
-    directly from the lightcurve FluxPoints object
-
-    Parameters
-    ----------
-    lightcurve : '~gammapy.estimators.FluxPoints'
-        the lightcurve object
-
-    Returns
-    -------
-    fpp: `~numpy.ndarray`
-        Point-to-point fractional variation
-
-    """
-
-    flux = lightcurve.flux.data.flatten()
-    flux_err = lightcurve.flux_err.data.flatten()
-
-    return compute_fpp(flux, flux_err)
-
-
-def eval_lc_timing(lightcurve, efolding=True):
-    r"""Wrapper to utilize the '~gammapy.stats.compute_etime'
-    and '~gammapy.stats.compute_2time' functions
-    directly from the lightcurve FluxPoints object
-
-    Parameters
-    ----------
-    lightcurve : '~gammapy.estimators.FluxPoints'
-        the lightcurve object
-    efolding : 'bool'
-        flag to compute e-folding time or doubling time
-
-    Returns
-    -------
-    t:  float
-        Doubling time or e-folding time
-    """
-    flux = lightcurve.flux.data.flatten()
-    time = lightcurve.geom.axes["time"].time_mid
-
-    if efolding:
-        return compute_etime(flux, time)
-    else:
-        return compute_2time(flux, time)
